requirements.txt

matplotlib~=3.8.4
dvclive~=3.46.0
kaggle~=1.6.12
os
seaborn~=0.13.2
numpy~=1.26.4
pandas==2.2.2
nltk==3.6.1
scipy==1.13
matplotlib==3.4.1
scikit-learn==1.4.2

Flask==1.1.4
markupsafe==2.0.1
jinja2<3.1.0
flasgger

pylint==3.1.0

tensorflow==2.16.1
tensorrt==10.0.1
keras==3.3.3
seaborn==0.13.2
joblib~=1.4.2
PyYAML~=6.0.1